Remove dependence of compute_globals on t_state
This change makes the different function related to compute_globals
request the fields of t_state explicitly, instead of simply taking
a pointer to the full t_state object. This simplifies refactoring
the t_state object in future commits. This change also clarifies why
both state->box and an additional parameter `matrix box` is required -
the latter one is the box at the last complete timestep in do_md and
has been renamed accordingly.

Note that in the documentation to checkNumberOfBondedInteractions(),
which was changed in a similar way, the argument `state` was
described as being the global state. Throughout the code, the local
state was used, however. This part of the documentation is now
changed, but if that was intended, then a follow-up change fixing
this might be necessary.

Change-Id: If913656fe9bf07567a29f68592402bb5ff17fc52