Get rid of clang-tidy warning in do_md/rerun/mimic
This change silences most of the clang-tidy warnings
"Clang-Tidy: Use of a signed integer operand with a binary
bitwise operator" in do_md, do_rerun and do_mimic. This warning
is triggered when using a signed integer with a binary bitwise
operator, as we do all over the code when encoding / decoding
flags. Such warnings are popping up in other parts of the code,
but this change takes only care of the ones which can be
silenced by changes in the simulator functions and the bit shift
definitions.

Change-Id: I7345e987e91051c9688cb2e8a96026682f025b6b