Fix pre-submit build matrix and description
Pre-submit should not fail because the ftp server is not reachable, so
moved it to nightly matrix. Also recorded the intent to test
such a build.

Also fixed that the intent to test AVX_128_FMA in mixed precision was
broken in Ib2c5e14e5d89665130ae8d6c749c3230ccfe72a9. There's no
ability to test this SIMD in pre-submit without wider changes, and their
is neither large risk nor consequence of breaking it, so nightly is OK.

Refs #2809, #2629

Change-Id: I9cd41a359ba662a8a9529540c50c9a7ef546cd29