Test specific hwloc versions
Now that releng understands the installed hwloc versions, we can
specify those in the matrix configurations. We can also specify that a
test configuration requires hwloc, but don't have to specify the
version. And we can leave it auto and provide a specific version of
the library to get detected automatically.

Refs #2899

Change-Id: I7f72b79746e77d04d9e2657b75821a8e3d62d83b