Modernize FindHwloc.cmake
Users (and Jenkins) need the ability to direct finding a particular
hwloc version, e.g. from a module. This is particularly important now
that two API versions exist and are supported in GROMACS. Imported the
find module from Inria, which adds several useful capababilities,
including the ability to specify where to look for a specific version
of this library.

Removed the pkgconfig support from the imported FindHwloc.cmake
because it seems the hwloc.pc file from libhwloc does not lead to the
CMake idiomatic behaviour of HWLOC_LIBRARIES being the full path to
the library.

Refs #2539, #3011

Change-Id: I6837144a556f1c61a0f13406fb31fe1d527b65f4