Constraint element for the modular simulator
This change introduces a constraint element, which is implemented for the
position-and-velocity and the velocity-only case. This change does not change
the constraint implementation itself, but replaces the `constrain_coordinates`
and `constrain_velocities` calls from update.h by elements implementing the
ISimulatorElement interface and using the new data management.

This commit is part of the commit chain introducing the new modular
simulator. Please see docs/doxygen/lib/modularsimulator.md for details
on the chosen approach. As the elements of the new simulator cannot all
be introduced in one commit, it might be worth to view Iaae1e205 to see
a working prototype of the approach.

Change-Id: Ie04b995fc34801967427825bb2842411a448a677