CompositeSimulatorElement for the modular simulator
This element allows to have a list of elements act as a single one.
This will simplify building algorithms.

This commit is part of the commit chain introducing the new modular
simulator. Please see docs/doxygen/lib/modularsimulator.md for details
on the chosen approach. As the elements of the new simulator cannot all
be introduced in one commit, it might be worth to view Iaae1e205 to see
a working prototype of the approach.

Change-Id: If2153923dd9421288ad57337c7255000bb7c1d91