Begin new sample_restraint development.
* Update the sample MD plugin project for its new home in the GROMACS
  source tree and the gmxapi Python package that is now part of the
  GROMACS distribution (as opposed to gmxapi 0.0.7 hosted at
  https://github.com/kassonlab/sample_restraint)
* Update googletest infrastructure to use CMake 3.9 utilities.
* Download googletest and pybind sources during docker CI build.
* Download binary test data from GitHub.

Future updates are necessary both for using the subtree in Jenkins
builds (to avoid downloading or using redundant copies of dependencies)
and for using the subtree as a starting point for an external project
(in which case the external dependencies _should_ be bundled).

Also, requires Change-Id: I44a66646eceb191a2ec755169671af326f22526a to
pass all CI tests.

Refs #3027

Change-Id: I4dd699168d6bc8e84e7de4d3b5b7b1cd6bd21090