Split canDetectGpus()
The function canDetectGpus() was used in a higher and lower level
fashion. The original canDetectGpus() is renamed to
isGpuDetectionFunctional(). Added canPerformGpuDetection() which also
checks for a GPU build and the env.var. GMX_DISABLE_GPU_DETECTION.

This also fixes an inconsistency error in the PME GPU unit tests
when running with GMX_DISABLE_GPU_DETECTION.

Change-Id: If762b2a108eedeb1bda0a2af9afa8a418aa93e14