Made parallel tune_fc faster by reducing communication.
Originally the whole poldata was broadcast from the
master node to all the slave nodes. Now only the parameter(s)
that change are being sent. For a small test case of
8 compounds the calculation is now faster on 2 or 4
processors than on 1 and faster on 4 than on 2, which was
not the case before.

TODO: Optimization of sequential code.
Change-Id: I7944cbefbad7a657180de56ec46bf7b929a70c92