Fix hwloc find package naming and quietness issues
Due to compound issues the hwloc detection has been more noisy at
subsequent cmake executions than it should.
This change corrects the module naming to "HWLOC" as this is the prefix
the variables the modules defines. Also implemented an internal cache
variable to detect subsequent calls to the hwloc detection.

Change-Id: I13c82277b1edf577ebdc4791d6cb2576c5ff434d