ForceElement and ShellFCElement for the modular simulator
These elements encapsulate the calls to do_force and relax_shell_flexcon,
respectively. Rewriting calls to force routines is not within the scope
of this effort, so this allows us not to worry about it.

This commit is part of the commit chain introducing the new modular
simulator. Please see docs/doxygen/lib/modularsimulator.md for details
on the chosen approach. As the elements of the new simulator cannot all
be introduced in one commit, it might be worth to view Iaae1e205 to see
a working prototype of the approach.

Change-Id: I0923ddfc9f8f477b8920d212bcbdcf641075f80b