Stop testing ICC on Windows
Windows is seldom used with GROMACS, and likely Windows+ICC is never
used. Maintaining the build system (including compiler flags) to work
with this combination, together with installing and updating ICC on
our Windows agent node is work that does not deliver value. We're not
likely to be able to afford to continue doing this in Gitlab Runner.

Moved the unrelated test coverage provided by the former post-submit
configuration to other configurations.

Updated install guide to reflect the new status (plus a few other
incidental updates).

Change-Id: I69868d29d95f959b31c1514f219d3d0940af00e6