Van der waals parameters are now optimized in tune_fc.
Fixed a bug where the van der waals parameters were not
propagated to the correct gromacs structure (forcerec->nbfp)
but only to mtop. That meant VdW params were essentially
drifting but were not optimized, in practice the initial
parameters were kept constant and the VdW energies did not
change despite that the parameters changed.

Change-Id: I67d1fbe6b4028b5d982f6a74dba68df6477893bb