Add sample_restraint Python tests to `check` target.
When building as part of the GROMACS project (not as a standalone
project), add a CMake testing target to invoke pytest for the
sample_restraint package. Include this target in the GROMACS project
`check` target. Make some minor CMake infrastructure changes to help
the sample_restraint tests find the gmxapi Python package.

Change-Id: I753951c1bb5f5a9c6caa20d4d632eba214895a1f