Activation of density fitting and move of module notifications
The densityfitting code is activated by adding it to the MDModules.

Moved the MdModuleNotifications to mdrunutility to break cyclic module
dependencies between mdrun and applied_forces via MdModules and
DensityFittingModuleInfo.

Observe that this patch changes the default .mdp output. Updated
reference data accordingly. Note that the changes in .mdp output
lead to "check mdp file differences" warning in the regression tests.

refs #2283

Change-Id: I93c0dd85285cac3f9e94d2dfc7c7fcc434aa81db