Fix issues for clang-analyzer-8
pdb2gmx can no longer attempt to add a vsite after issuing a warning
that it cannot convert an atom, instead exiting with a fatal error.

Fixed bug in gmx order -calcdist

Suppressed some analyzer leak warnings of gmock objects. If the leaks
are real, ASAN will really find it, and if not we aren't going to fix
it in gmock.

Change-Id: Ib3290d7722c1e719ce5291cf3db68a87bc7ce89a