Work around clang CUDA device code codegen bug
Some of the PME kernels (namely PME spread and solve) get miscompiled
when clang native device code compilation is used with assertions and
optimization on, causing errors in RelWithAssert builds.
As a workaround, this change disables optimization for these kernels when
assertions are enabled (and only if the device compiler is clang).

Refs #2819

Change-Id: I815fc86116ffb57c4d5803ce7fa4d260909ae7ba