v-rescale thermostat for modular simulator
This commit introduces the v-rescale thermostat for the modular
simulator. It also implements general velocity scaling in the
propagators.

Note that since the modular simulator is the default code path for
all supported input using md-vv as an integrator, this automatically
moves over some of the test coverage from the legacy simulator to
the modular simulator. This does, however, also extend equivalence
testing between the modular and the legacy code paths to v-rescale
thermostatted simulations.

Change-Id: Ia22931ccc3f9340f7f6272e395df149bd73d9b21