Allow only one EEM model in poldata files.
Force field parameters are interdependent. Therefore it is necessary
to co-develop and optimize them. Here the possibility to list multiple
EEM parameter sets in a poldata file (previously gentop.dat) has been
removed. Instead, there is a specific file for each of the models
(for now AXg and AXpg only). The gentop module is aware of this
and will find the right force field file if presented with the
-qdist option.

Change-Id: I539a72906138b8fa2275b923532287ff947ab2cd