Extend GPU traits class
Now GPU traits provide a non-GPU header, so that generic code
can use CommandStream, CommandEvent and DeviceContext types.
The header also diverges to a platform-specific version when
needed upon compilation. This change allows for passing the
variables of the above types in the general (non-GPU) parts
of the code and can be included where the code is shared
between different platforms.

Renamed a Context variable to DeviceContext for greater clarity.

Change-Id: If21b9dacac66ff7203948eb03de96f9473b7359a