Renamed all the force fields and removed options from gentop.
The concept of two flags -qdist and -qgen was removed since it
is not possible to mix and match for a user. Only supported
force fields are left to be used. For gentop that means for now
ACM-g and ACM-pg are available to generate whole topologies
whereas a numnber of "legacy" force field are available to
generate charges only.

Because of renaming variables a lot of code is touched unfortunately.

Change-Id: I2dc0ce8315765f906d2ab57f80c454b536d02bd0