Update mdrun test comparison infrastructure
This change is refactoring and extension, no use is made of any new
capabilities.

Referred to position coordinates as coordinates, in line with
agreement elsewhere. Also referred to energy terms consistently,
rather than energy fields or tolerances.

Extended the range of mdp option parameters able to be tested
configurably with the contents of the simulation database.

Made energy and trajectory comparison machinery easier to use.

Trajectory comparison is now more configurable. Sometimes the test
runner will know that it is OK to compare the data in a test frame
with a reference frame even though one of them has data that the other
does not have, or will differ, etc. So, it is now possible to compare
e.g. a test trajectory that has a frame with velocities and not forces
with a reference trajectory whose frame has both velocities and
forces, and have the comparison be valid or not according to whether
the test runner expected the mismatch in contents.

Fixed issue with trajectory frame comparison assuming that coordinates
would always be present.

More use of std::string.

Change-Id: I6926380e271344e7bd1a7c6594cc604d7a68e67a