Adapt inheritance intendation to uncrustify
As discussed in I35f310da and some other commits along the modular
simulator chain, uncrustify insists on aligning the first inheritance
with the `final` keyword, as in

template <ComputeGlobalsAlgorithm algorithm>
class ComputeGlobalsElement final :
    public                  ISimulatorElement,
    public IEnergySignallerClient,
    public ITrajectorySignallerClient,
    public ITopologyHolderClient

It was therefore decided to align all inheritances to the same level,
as in

template <ComputeGlobalsAlgorithm algorithm>
class ComputeGlobalsElement final :
    public                  ISimulatorElement,
    public                  IEnergySignallerClient,
    public                  ITrajectorySignallerClient,
    public                  ITopologyHolderClient

This formatting will be reevaluated when uncrustify is replaced by
clang-format.

This commit applies this decision to the classes previously introduced
along the commit chain. It also addresses two typos in trajectoryelement.h.

Change-Id: I03b9fd142d822f7a7523e351f08a4044bff2cff4