Rename MicroState to StatePropagatorData for modular simulator
The name of the t_state replacement is changed from MicroState
to StatePropagatorData. Also, the write and read access to x/v/f
data is renamed to xxxView() and constXxxView() for additional
clarity and to avoid confusion with file writing operations.

This also removes the constraint element from the list of friend
classes, as there is no need for direct access to the state
anymore thanks to I28b58df4.

Change-Id: I1b5b966aafbfa6fe5c44c36416ebb25f38505b58