Add gpuupdate and gpucomm OnDemand tokens
These allow triggering the GPU update-constraints and the GPU direct
communcation development features (both PP-PME and halo exchange),
respectively.

Also added matrix configs for both with configs identical to those in
the gpubufferops for the sake of simplicity.

Added an OpenCL config to ensure the additons of uninmpled features on
that codepath don't break things.

Change-Id: Iaf308d9b55ff10d8e43b655c2e1708eb05a5b43c