Basic toolchain to facilitate GMXAPI build
Adds a basic toolchain file and install directives for it.
This is aimed at having one point for the API to look up how
a native GROMACS installation was build and allow the bindings for the
API to be build in the same way.

Added hint for directory to GMXRC so that it will be easier to find.

Refs #2961

Change-Id: Icefe89e97009110be55dc8e1f3db5726ec1fe53a