Add basic interface to run update on GPU
Added basic command line option to direct update on a GPU,
as well as adding the necessary checks.

Since now the default code path is moved to the GPU, overly
tight tolerances in simulator tests lead to the failure of
one of the tests. This was a result of a slightly low
absolute value of the kinetic energy specified in the
definition of  tolerances, which is also corrected by this
patch.

Change-Id: Ieca885d935aa1e0d592735bc7027f8261288b635