StatePropagatorData: Introduced const box accessor and fixed legacy flags
Two improvements for the StatePropagatorData:

The current implementation only gave access to the non-const box
matrix. This introduces a const version of it, and changes the elements
using the simulation box to use the const version where appropriate.

The current implementation also kept using the legacy flags it got
from the global_state in all further usages. This is not correct -
when StatePropagatorData creates a deep copy as a t_state object, it
only populates certain fields. This commit fixes this.

Change-Id: I783491d9f8a0825b169880b7ab8ef7e677623d50