Change PaddedVector to PaddedHostVector for force CPU buffer
The difference between PaddedVector and PaddedHostVector is that
in the builds where CUDA is available, cudaMalloc is used instead
of malloc. This enables one to pin the vector for asynchronous
H2D and D2H copies. Other than that and on the CPU-only builds this
change essentially only renames the data type.

Change-Id: I446ef037a8013f666a78762d37ed44e59f9548c2