Refactor to create a GpuTaskAssignments object
This change is purely reorganization, nothing is reordered or
functionality changed.

This expands on the previous vector of vectors of task assignments
into an object with the invariant of a completed assignment. It has
very complex construction logic, so it has a separate builder class
for the data that describes the decision about which GPUs are assigned
to tasks running on ranks on each physical node.

This refactoring prepares for re-ordering mdrun setup
so that the decision of rank duty by DD can lead to
task assignment, device initialization and the creation
of GPU contexts and streams with which to initialize
the various modules that use them to do their work.

Minor improvements to documentation and naming in a few places.

Change-Id: Idec71a3016212a4125b23b780e1059dbb134d7fb