Fix domdec for md-vv modular simulator
Domdec might have changed the local velocity vector after the state
object made a backup to restore after the first md-vv half step.
This change makes sure that the backup is updated if this happens.

Note that the half-step resetting will be eliminated in the course
of the transition to the modular simulator, but is currently kept
for backwards compatibility.

Change-Id: Ifb6346055a14d34a4efcae43b7b41399b1444bca