Fix vrescale thermostat integral (modular simulator)
The v-rescale thermostat was calculating the thermostat integral,
but this was neither written out nor saved into checkpoints. The
current commit fixes this.

Note that this is a good example of why the energy element needs
to become more modular - it should not need to know about the
details of other elements. This is a todo moving forward though,
not something we can solve immediately now.

Change-Id: I69541cb8ffa3c07e294e583f2c20423952e3ec01