Fix the bug with offloading the PME-only ranks to GPU
This fixes the bug introduced with StatePropagatorDataGpu (Commit 092a8f68):

In PME-only ranks, the GPU context and stream were mixed up and passed to
the constructor of StatePropagatorDataGpu in the wrong order. This caused
failure in OpenCL builds with error -34 (CL_INVALID_CONTEXT). Did not
affect the CUDA builds, since the context is not relevant in CUDA.

Change-Id: I5e070e361c7bdef3168887dcac4869bb71c6c5ed