Fix random energy and virial with PME on GPU on a PME-only rank
Missing zero initialization of never computed LJ PME energy and virial
terms with PME on the GPU could lead to random energy, virial and
pressure numbers.

The effect of this bug was that the potential and total energy could
be off (not the Coulomb mesh energy). This didn't affect sampling.
The pressure could be off, which would affect sampling when pressure
coupling is used, but likely the simulation would explode after
a few steps.

Fixes #3120

Change-Id: I309dde958f1b73e7f71f87f4f5ad016d16f8d16b